feat: fallback repository implementation
